requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
tree-sitter==0.21.3
tree-sitter-python==0.21.0
tree-sitter-java==0.21.0
tree-sitter-javascript==0.21.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
rank-bm25==0.2.2
networkx==3.3
openai==1.40.0
pydantic==2.8.2
pydantic-settings==2.4.0
gitpython==3.1.43
PyYAML==6.0.2
markdown==3.6
tiktoken==0.7.0
loguru==0.7.2
httpx==0.27.0